+++ Assets/scripts/Objects/Armour.cs
@@ -152,7 +152,7 @@ public class Armour : Equipment
 
     public override int repairEstimate()
     {
-        return (short)(GameWorldController.instance.objDat.armourStats[item_id - 32].durability);
+        return GameWorldController.instance.objDat.armourStats[item_id - 32].durability;
     }
 
 }
