+++ Assets/scripts/Traps/a_change_from_trap.cs
@@ -24,7 +24,7 @@
             return;
         }
 
-        short NewTileFloorTexture = (short)((int)ChangeTo.heading | ((((int)ChangeTo.zpos >> 4) & 0x1) << 3));
+        short NewTileFloorTexture = (short)((int)ChangeTo.heading | (((ChangeTo.zpos >> 4) & 0x1) << 3));
 
         short tileFloorCriteria = (short)((int)heading | (((zpos >> 4) & 0x1) << 3));
 
@@ -53,7 +53,7 @@
                     CurrentTileMap().Tiles[x, y].floorTexture = NewTileFloorTexture;//ChangeTo.heading;
                     if (ChangeTo.owner < 10)
                     {
-                        CurrentTileMap().Tiles[x, y].tileType = (short)ChangeTo.owner;
+                        CurrentTileMap().Tiles[x, y].tileType = ChangeTo.owner;
                     }
                     if (ChangeTo.zpos < 15)
                     {
